Extract View extension into its own file
